+++ R/ts_examples.R
@@ -52,7 +52,7 @@ ts_dygraphs <- ts_(dygraphs::dygraph, class = "xts", reclass = FALSE)
 
 #' @export
 #' @name ts_examples
-ts_forecast <- ts_(function(x) forecast::forecast(x)$mean, vectorize = TRUE)
+ts_forecast <- ts_(function(x) forecast::forecast(ts_na_omit(x))$mean, vectorize = TRUE)
 
 #' @export
 #' @name ts_examples