+++ tests/testthat/test_ts_span.R
@@ -67,4 +67,28 @@ test_that("works with non-heuristic frequencies (#106)", {
 
 
 
+test_that("extending by NA works", {
+  expect_equal(
+    ts_span(mdeaths, end = "19801201", extend = TRUE),
+    window(mdeaths, end = c(1980, 12), extend = TRUE)
+  )
+
+  expect_equal(
+    ts_span(mdeaths, start = "1973", extend = TRUE),
+    window(mdeaths, start = 1973, extend = TRUE)
+  )
+
+})
+
+test_that("extending by NA works for different frequencies", {
+  x <- ts_span(ts_df(ts_c(AirPassengers, austres)), start = 1930, extend = TRUE)
+  expect_identical(
+    ts_span(x, start = "1930-01-01", end = "1930-01-01")$value,
+    c(NA_real_, NA_real_)
+  )
+})
+
+
+
+